# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginReturnEntity" }
              }
            }
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserEntity" }
              }
            }
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginEntity" }
              }
            }
          },
          "401": {
            "description": "token is not valid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HttpExceptionEntity" }
              }
            }
          },
          "500": {
            "description": "generate token exception",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HttpExceptionEntity" }
              }
            }
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/github/login": {
      "get": {
        "operationId": "AuthController_githubLogin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "string" } }
            }
          },
          "401": { "description": "" }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Auth"]
      }
    },
    "/auth/github/callback": {
      "get": {
        "operationId": "AuthController_githubCallback",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Auth"]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Filter notifications by projectId",
            "schema": { "example": 1, "type": "number" }
          },
          {
            "name": "isSeen",
            "required": false,
            "in": "query",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPaginateEntity"
                }
              }
            }
          }
        },
        "summary": "Retrieve a list of notifications (filter by projectId & paginate)",
        "tags": ["notifications"]
      }
    },
    "/notifications/{id}": {
      "patch": {
        "operationId": "NotificationsController_updateNotification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateNotificationDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["notifications"]
      }
    }
  },
  "info": {
    "title": "THESIS",
    "description": "The API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password"
          }
        },
        "required": ["email", "password"]
      },
      "RoleEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "category": {
            "type": "string",
            "enum": ["MEMBER", "GUEST", "OWNER"]
          },
          "name": { "type": "string" },
          "projectId": { "type": "number" }
        },
        "required": ["id", "category", "name", "projectId"]
      },
      "UserEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "email": { "type": "string" },
          "fullname": { "type": "string" },
          "avatar": { "type": "string" },
          "address": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "role": { "$ref": "#/components/schemas/RoleEntity" },
          "phoneNumber": { "type": "string" },
          "username": { "type": "string" },
          "githubId": { "type": "string" },
          "githubUsername": { "type": "string" }
        },
        "required": [
          "id",
          "email",
          "fullname",
          "createdAt",
          "phoneNumber",
          "username"
        ]
      },
      "LoginReturnEntity": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "user": { "$ref": "#/components/schemas/UserEntity" }
        },
        "required": ["accessToken", "refreshToken", "user"]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "email": { "type": "string" },
          "phoneNumber": { "type": "string" },
          "address": { "type": "string" },
          "password": { "type": "string" },
          "fullname": { "type": "string" },
          "avatar": { "type": "string" }
        },
        "required": ["username", "email", "phoneNumber", "password", "fullname"]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Refresh Token",
            "example": "d50a44bc-5249-4406-83b1-afdea1887243"
          }
        },
        "required": ["token"]
      },
      "LoginEntity": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "user": { "$ref": "#/components/schemas/UserEntity" }
        },
        "required": ["accessToken", "refreshToken"]
      },
      "HttpExceptionEntity": { "type": "object", "properties": {} },
...
      "NotiFieldInformation": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "number" }
        },
        "required": ["name"]
      },
      "NotificationContent": {
        "type": "object",
        "properties": {
          "subject": { "$ref": "#/components/schemas/NotiFieldInformation" },
          "action": {
            "type": "string",
            "enum": ["CREATE", "UPDATE", "DELETE"]
          },
          "objectEntity": { "type": "string", "enum": ["REPORT", "TASK"] },
          "object": { "$ref": "#/components/schemas/NotiFieldInformation" },
          "objectAttribute": { "type": "string" },
          "objectAttributeValue": {
            "$ref": "#/components/schemas/NotiFieldInformation"
          }
        },
        "required": ["subject", "action", "objectEntity", "object"]
      },
      "NotificationEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "content": { "$ref": "#/components/schemas/NotificationContent" },
          "createdAt": { "format": "date-time", "type": "string" },
          "projectId": { "type": "number" },
          "userId": { "type": "number" },
          "isSeen": { "type": "boolean" },
          "ProjectMember": { "type": "object" }
        },
        "required": [
          "id",
          "content",
          "createdAt",
          "projectId",
          "userId",
          "isSeen",
          "ProjectMember"
        ]
      },
      "NotificationPaginateEntity": {
        "type": "object",
        "properties": {
          "total": { "type": "number" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/NotificationEntity" }
          }
        },
        "required": ["total", "items"]
      },
      "UpdateNotificationDto": {
        "type": "object",
        "properties": { "isSeen": { "type": "boolean" } },
        "required": ["isSeen"]
      }
    }
  },
  "security": [{ "bearer": [] }]
}